Moved simulated node to SDK and capacity connector and map to capacity.
